chore: Remove support for extension field without chip
Will add support back later at a higher level that is easier to maintain
